# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.HTTPError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  types.Todo:
    properties:
      completed:
        type: boolean
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is a todo list management application
  license: {}
  title: Todo Application
  version: "1.0"
paths:
  /todos:
    get:
      consumes:
      - application/json
      description: Get all todo items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Todo'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Get all todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo item
      parameters:
      - description: New Todo
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/types.Todo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Create a todo
      tags:
      - todos
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a new todo item
      operationId: get-string-by-int
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/types.Todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Delete a todo
      tags:
      - todos
    get:
      consumes:
      - application/json
      description: Get a todo item
      operationId: get-string-by-int
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Todo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Get a todo
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Update a todo item
      operationId: get-string-by-int
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Todo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.HTTPError'
      summary: Update a todo
      tags:
      - todos
swagger: "2.0"